Update README.md
Write a program, with comments, to do the following: 
Ask the user to enter a string with at least 10 characters assign it to variable inp.
a. If inp has less that 10 characters, print “Invalid input!” and exit.
b. Otherwise, if inp is an alphabetic string,
i. print the las 4 characters of inp in uppercase.
ii. check how many times the letter ‘e’ occurs in inp and print this value with
a suitable message.
c. Otherwise, print “You have entered a valid non-alphabetic string.”
Some sample outputs for the 3 questions are given below.